Refactor file structure and rename components
