+++ x-pack/plugins/upgrade_assistant/common/constants.ts
@@ -28,5 +28,6 @@ export const indexSettingDeprecations = {
 
 export const API_BASE_PATH = '/api/upgrade_assistant';
 
+export const DEPRECATION_WARNING_UPPER_LIMIT = 999999;
 export const DEPRECATION_LOGS_SOURCE_ID = 'deprecation_logs';
 export const DEPRECATION_LOGS_INDEX_PATTERN = '.logs-deprecation.elasticsearch-default';
