+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/identify_step/_index.scss
