+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/identify_step/deprecation_logging_toggle/_deprecation_logging_toggle.scss
