+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/identify_step/external_links.tsx
