--- x-pack/plugins/upgrade_assistant/public/application/components/overview/resolve_step/_index.scss
@@ -1,1 +1,0 @@
-@import 'no_deprecations';
