+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/review_step/_index.scss
@@ -1,0 +1,1 @@
+@import 'no_deprecations/no_deprecations';
