+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/review_step/es_stats/es_stats_error.tsx
@@ -8,8 +8,8 @@
 import React from 'react';
 
 import { EuiIconTip } from '@elastic/eui';
-import { ResponseError } from '../../../lib/api';
-import { getEsDeprecationError } from '../../../lib/es_deprecation_errors';
+import { ResponseError } from '../../../../lib/api';
+import { getEsDeprecationError } from '../../../../lib/es_deprecation_errors';
 
 interface Props {
   error: ResponseError;
