+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/review_step/kibana_stats/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { getResolveStep } from './resolve_step';
+export { KibanaDeprecationStats } from './kibana_stats';
