+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/review_step/no_deprecations/_no_deprecations.scss
