+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/review_step/no_deprecations/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { getObserveStep } from './observe_step';
+export { NoDeprecations } from './no_deprecations';
